+++ apk/utils/__init__.py
@@ -1,1 +1,2 @@
 from .markdown_defs import *
+from .net_utils import *
