+++ apk/utils/net_utils.py
@@ -1,16 +1,6 @@
-import re
 import subprocess
 
 from jnius import autoclass
-
-
-def is_valid_mac(mac):
-    return re.match(r'^([0-9A-Fa-f]{2}[:-]){5}([0-9A-Fa-f]{2})$', mac)
-
-
-def mac2ipv6_ll(mac, pref):
-    m = hex(int(mac.translate(str.maketrans('', '', ' .:-')), 16) ^ 0x020000000000)[2:]
-    return f'{pref}::%s:%sff:fe%s:%s' % (m[:4], m[4:6], m[6:8], m[8:12])
 
 
 def get_ipv6_with_su(iface):
