+++ utils/__init__.py
@@ -1,6 +1,5 @@
 from .argparser import *
 from .defines import *
-from .os_utils import *
-from .net_utils import *
+from .misc_utils import *
 from .output_manager import *
